refactor: Edit Config class and Server class
- Remove unused code lines from Config class and Server class.
- Refactor Server::start() method into pieces of function blocks.
- Separate and detach socket and network related code block from Server.cpp to new file Server-network.cpp file.
- Remove unused class Poller and Socket related code and files.
- Update webserv.hpp file and Makefile, with the above changes.
